requirements.txt
asgiref==3.2.3
astroid==2.3.3
backcall==0.1.0
decorator==4.4.1
Django==3.0.7
entrypoints==0.3
flake8==3.7.9
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pytz==2019.3
djangorestframework==3.11.0
six==1.14.0
snowballstemmer==2.0.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.8
wrapt==1.11.2
